view model setup

+++ app/src/main/java/com/example/forecast/data/db/entity/Current.kt
@@ -16,7 +16,7 @@ data class  Current(
     val precip: Double,
     val pressure: Double,
     val temperature: Double,
-    val visibility: Int,
+    val visibility: Double,
     @SerializedName("weather_code")
     val weatherCode: Int,
     @SerializedName("weather_descriptions")
